[build] optimize make fmt
Now there are separate `fmt-sk`, `fmt-c`, and `fmt-js` targets that
run `skfmt`, `clang-format`, and `prettier`. The `skfmt` and
`clang-format` invocations are parallelized, but `prettier` is not
since its startup time is so high that running in parallel is even
slower. Also note that `make -j fmt` runs all 3 targets in parallel.