[build] Add git pre-commit hook to check formatting
Install locally by running `make setup-git-hooks`. At that point,
running `git commit` will invoke the pre-commit script to check that
the staged changes are well-formatted. If so, it silently succeeds and
`git commit` proceeds as before. If not, it reports the files that
changed and the diff between the staged code and the formatted
code. At this point `make -sj fmt` will format the files in the
working directory, and the changes can be staged and the commit
process continued.

Note that to override the check and commit even if it fails, use `git
commit --no-verify`.