Force a rebuild when a client reconnects to a new server
Why? Because a new server loses watermarks and has a different tick.
This breaks all the metadata that replication relies on. The rebuild
mechanism is fundamentally: reset your view of the server's clock and
resync your data to this. So it is a perfect fit to solve this
problem.

I disregarded trying to port the watermark data and initialise a new
server with the same tick as the old. We would also have to dump and
import all the Source data for each and every row. This is a much
bigger lift and all becomes irrelevant with vector clocks.

To actually trigger a rebuild the tailer maintains a new non-final
state in the replication state machine. Some methods were renamed to
make it clear this is a more general state machine now.