Evaluate side effect expressions separately
We do this so that nested expressions behave as expected, i.e. we need
to separate checks from params as they cannot be selected and are
different concepts - just syntatically similar.

We leave ASuccess, although not strictly needed now, so that CHECK
expressions inside nested expressions fail. Otherwise we will evaluate
the inner expression ignoring the CHECK aspect which is probably quite
confusing for a reader.