Rework can-mirror to locally check schemas and fail fast
`can-mirror` now explicitly compares a remote table/schema against the provided
schema, instead of relying on the existing local table's schema.

The previous behavior delayed/obscured failure in some cases, causing the client
to hang since failures were happening on the remote.  Now, there should be a
local error if a user attempts to mirror a table with an incorrect or
incompatible schema.