Fix eval loop segfault
Segfault happened on any second error. Basic repro:

```
skdb> select 1 + check(0);
select 1 + check(0);
^
|
 ----- ERROR
Error: line 1, character 0:
Failed constraint: 0
skdb> select 1 + check(1);
[1]    36873 segmentation fault  ./build/skdb
```

Tracking this down it was happening while looking in to the
context.globals sorted map during the printContext function called as
part of `SKIP_call_after_unlock`.

The manual memory management, while holding on to the globals object,
in _exec looked pretty suspect. My hunch is that we end up referencing
something on an obstack that is destroyed. So I refactored to copy out
the exn and handle this higher in the stack. Seems to work well.