Show possible fields on invalid "." (#249)
* Show possible fields on invalid "."

If you write `a.b` and `b` is a field that does not exist. You will have
a suggestion like "did you mean ...". This is useful, but that
suggestion is not always correct. This diff keeps that suggestion, but
also list all the fields that are accessible, directly in the error
message.

* Removed synonyms from the list of proposed field name options on error

The error message listing the available options for field names was
including synonyms.

* Fixed unit tests that was still proposing synonym fields in the error message