[std] Make `spawnp` errors recoverable.
This changes the return type of `Posix.spawnp()` (as well as that of
`Posix.Popen::create()`, `Posix.popen()`, and `System.subprocess()`)
to a `Result<_, IO.Error>` instead of having the runtime `spawnp()`
function to call `exit()` (for instance when attempting to execute a
binary that cannot be found in the current path).

The call sites are updated to maintain current behavior (that is,
failing upon `Failure()` being returned).