Refactor iterator materialization for consistent results
Previously, this only materialized when `clone()` is called, which could lead to
situations where a partially-traversed iterator gets cloned and produces a
partial clone.  Now, whenever we traverse past the first element of an iterator,
we'll eagerly splat it out into memory so that it can be re-traversed
predictably in the future.