[skstore] Allow mappers to close only over handles
This diff removes the class type argument from map functions in the
SKStore API, and changes the parameters for the Mapper class
constructors from effectively
```
...params: any[]
```
to
```
eparams: EHandle<any, any>[],
lparams: LHandle<any, any>[]
```

The benefit of these changes is that calls to the `map` functions no
longer need to specify the type of the class. This helps since type
inference is not strong enough to elaborate the type parameters
automatically. Also, specifying the types of each parameter as only
either `EHandle` or `LHandle` gives API users help to only close over
state that the SKStore incremental computation engine is aware of.

A drawback is that splitting the params into a pair of arrays rather
than one varargs list is less convenient at call sites. Additionally,
the `constructor` functions of mapper classes have a uniform
signature that is only very weakly typed: the number and order of the
actual class constructor parameters must be ensured manually.