[skc] Introduce `#env()` macro.
The `#env()` macro takes a single string literal as parameter, and
expands to the value of the corresponding environment variable at
compile time. It plays a somewhat similar role as the `-D` option of
gcc in that it allows injecting values into the program from the
command line.

This will make it easier to handle things such as embedding the git
commit hash of a given build into programs and displaying it with a
`--version` option.

Note: The environment is frozen when building an `sklib`, i.e. a call
to `#env("foo")` within a package `pkg` will always expand to the
value of the `foo` environment variable at the time of compilation of
`pkg`, even when `libpkg.sklib` is used as a dependency of an other
program/library.