Add target to check ts sources
This adds a `make check-ts` target to run the type-checker in all
directories in the npm workspaces list. Executing `npm run build` already
does run the type-checker in all such directories, but does so in a way that
causes `tsc` diagnostics to have locations with file names relative to the
subdirectory containing the `tsconfig.json` file, not the repo root, and
without any intervening output indicating which directories are entered and
exited as execution proceeds. This PR's script adds such output so that
locations in diagnostics can be automatically jumped to.