Fixes #496: support for joinCollections
This diff implements the 'join' operation of two collections.
It is assumed that:
- both collection values are objects (keys can be of any type)
- there never is more than 1 object on boths sides at any time
  unless the option allowNN is passed (which should be rare).

If both those conditions are met, given a key, if the first collection
defines an object of type V1 and the second collection defines an
object V2 for the same key, the resulting collection will associate
{...V1, ...V2} for that key. No value is associated if either side is
missing.