Strengthen typing of SkipService.initialData vs createGraph
This diff changes the type parameters of SkipService so that they only
specify corresponding collection names, key types and value types, without
already instantiating the EagerCollection type. This type "signature" can
then be used to construct the instantiations of Entry for the type of
initialData as well as the instantiations of EagerCollection for the type of
e.g. createGraph.

This uses typescript's "mapped" and "indexed access" types but does not need
the machinery of "conditional" or "infer" types.