Fix build failures due to zombie symlinks
Before this PR the following reliably fails for me:
```
$ npm install && npm run build
$ git clean -Xdf
$ npm install && npm run build
```
In the final step, the build sees the symlinks created in the first step
that were deleted in the second step as some sort of zombie, where
`fs.exists` thinks they do not exist but `fs.symlink` fails because they
do. Checking in a shell shows that `ls` is similarly confused.

This PR changes the build from checking if the link exists and only if not
calling symlink to creating a symlink with a temporary name and then
renaming it to the desired name. The temporary names are randomly generated,
and POSIX requires the rename to be atomic.