Changing the order in which we update sub-directories.
We use to have a scheme that was pretty simple. Update the directories
in their order of creation. The scheme works, but the problem is that
this order is sub-optimal when the topology of subdir changes.

Imagine the following scenario:

```
creating dir A/ which is empty
creating dir B/ which reads A

A gets updated, and now creates subdirs.
```

From this point on, the subdirs in A will always be updated **after**
B. Which means that, more often than not, B will be computed twice.

To fix this, this diff introduces the notion of a TimeStack. The order
in which the directories are computed is no longer based on just
creation time, but the stack of all the creation time of all the
creators.

So if directory A, creates a subdir S, the timeStack will be
```Array[A.creationTime, S.creationTime]```.

This way, we guarantee that the subdirs are always updated before the
next directories at the same level as the parent.