Simplify LazyDir#update
Instead of collecting all the (dirName, key) pairs in a set,
iterate on the original map directly.

Paths are exactly (dirName, key) pairs and sorted identically, we
won't have duplicates doing like that.

Not sure the setDir is still useful in case nothing is changed, but it
was here and it can't hurt (but insignificantly performance)