Simplify accumulation of dirties
Before dirties were accumulated in a list before being converted to a map.
Each updateWithStatus loop iteration would bring the new dirties to the map
and clear the list of new dirties.

Now, we put them directly in a map, that we keep till the end of updateWithStatus.

Unless update is reentrant, the behaviour should be the same.